Snap: Make sure photos get saved when app exits
Although we have a media save service that runs in the background, it's
not useful because the JPEG callback function runs in the UI thread, which
gets killed when the app exits. When this happens, pictures that have not
been processed through the JPEG callback before the app exits are
permanently lost. To reproduce, simply capture a photo and leave the app
before the photo is saved; this will cause the captured photo to never be
saved.

To fix this, place the takePicture() callback routines into a separate
thread and stall the UI thread until all photos are saved when the app
exits. This solution is messy, but it works.

Change-Id: Iccb64c3306ed4b8beca3b50e7ed5f393ac8b8406
Signed-off-by: Sultanxda <sultanxda@gmail.com>