Snap: Always perform autofocus before capture when using continuous AF
The OP3(T)'s continuous autofocus configuration is very buggy and
constantly throws the camera out of focus. This causes many pictures taken
to come out blurry and out of focus. This bug is in the BSP, so we cannot
truly fix it.

To improve the focus situation, force an autofocus event every time a
photo is captured while continuous autofocus is in use.

Change-Id: I5cf35cb8835cb7cb061fec185cd66c1886d2dfc4